requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
meld3==1.0.0
paho-mqtt==1.0
pyserial==2.7
requests==2.4.3
supervisor==3.1.2
tornado==4.0.2
wsgiref==0.1.2
